+++ packages/react-scripts/config/webpack.config.js
@@ -40,6 +40,17 @@ const getCacheIdentifier = require('react-dev-utils/getCacheIdentifier');
 // @remove-on-eject-end
 const postcssNormalize = require('postcss-normalize');
 
+const threadLoader = require('thread-loader');
+
+// This boots the max number of workers in the pool and loads specified modules into the node.js module cache.
+const jsWorkerPool = {
+  // can be set to Infinity for watching builds to keep workers alive
+  poolTimeout: 2000,
+};
+
+// To prevent the high delay when booting workers it possible to warmup the worker pool.
+threadLoader.warmup(jsWorkerPool, ['babel-loader']);
+
 const appPackageJson = require(paths.appPackageJson);
 
 const sassFunctions = require('bpk-mixins/sass-functions');
@@ -472,6 +483,10 @@ module.exports = function (webpackEnv) {
                 ...customModuleRegexes,
               ],
               use: [
+                {
+                  loader: require.resolve('thread-loader'),
+                  options: jsWorkerPool,
+                },
                 isEnvDevelopment && {
                   loader: require.resolve('cache-loader'),
                   options: {
@@ -545,6 +560,10 @@ module.exports = function (webpackEnv) {
               test: /\.(js|mjs)$/,
               exclude: /@babel(?:\/|\\{1,2})runtime/,
               use: [
+                {
+                  loader: require.resolve('thread-loader'),
+                  options: jsWorkerPool,
+                },
                 isEnvDevelopment && {
                   loader: require.resolve('cache-loader'),
                   options: {