+++ packages/react-scripts/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 9.1.0 - UNRELEASED
 
-Add `cache-loader` to speed up local development
+- Add `thread-loader` to speed up webpack workflow by bundling in parallels
+- Add `cache-loader` to speed up local development
 
 ## 9.0.3